+++ src/routes/dashboardPages/inscricao/comoConheceu/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ComoConheceu from './ComoConheceu';
+
+export default {
+
+  path: '/inscricao1',
+
+  action() {
+    return <ComoConheceu />;
+  },
+
+};
