+++ src/routes/dashboardPages/inscricao/criarSenha/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import CriarSenha from './CriarSenha';
+
+export default {
+
+  path: '/inscricao4',
+
+  action() {
+    return <CriarSenha />;
+  },
+
+};
