+++ src/routes/dashboardPages/inscricao/dadosPessoais/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import DadosPessoais from './DadosPessoais';
+
+export default {
+
+  path: '/inscricao2',
+
+  action() {
+    return <DadosPessoais />;
+  },
+
+};
