+++ src/routes/dashboardPages/inscricao/endereco/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Endereco from './Endereco';
+
+export default {
+
+  path: '/inscricao3',
+
+  action() {
+    return <Endereco />;
+  },
+
+};
